+++ packages/model-typings/src/models/IReadReceiptsModel.ts
@@ -1,5 +1,5 @@
-import type { IReadReceipt, IUser, IMessage } from '@rocket.chat/core-typings';
-import type { FindCursor, DeleteResult, UpdateResult, Document, Filter } from 'mongodb';
+import type { IReadReceipt } from '@rocket.chat/core-typings';
+import type { FindCursor, DeleteResult } from 'mongodb';
 
 import type { IBaseModel } from './IBaseModel';
 
@@ -10,14 +10,5 @@ export interface IReadReceiptsModel extends IBaseModel<IReadReceipt> {
 	removeByRoomIds(roomIds: string[]): Promise<DeleteResult>;
 	removeByMessageId(messageId: string): Promise<DeleteResult>;
 	removeByMessageIds(messageIds: string[]): Promise<DeleteResult>;
-	removeByIdPinnedTimestampLimitAndUsers(
-		roomId: string,
-		ignorePinned: boolean,
-		ignoreDiscussion: boolean,
-		ts: Filter<IMessage>['ts'],
-		users: IUser['_id'][],
-		ignoreThreads: boolean,
-	): Promise<DeleteResult>;
-	setPinnedByMessageId(messageId: string, pinned?: boolean): Promise<Document | UpdateResult>;
-	setAsThreadById(messageId: string): Promise<Document | UpdateResult>;
+	findOlderThan(date: Date): FindCursor<IReadReceipt>;
 }