+++ packages/ddp-client/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rocket.chat/ddp-client
 
+## 1.0.8
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@8.3.2
+  - @rocket.chat/rest-typings@8.3.2
+  - @rocket.chat/api-client@0.2.55
+  </details>
+
+## 1.0.7
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@8.3.1
+  - @rocket.chat/rest-typings@8.3.1
+  - @rocket.chat/api-client@0.2.54
+  </details>
+
 ## 1.0.6
 
 ### Patch Changes