+++ packages/omni-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rocket.chat/omni-core
 
+## 0.0.20
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/models@2.1.4
+  </details>
+
+## 0.0.19
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/models@2.1.3
+  </details>
+
 ## 0.0.18
 
 ### Patch Changes